+++ app/_components/FlexRow/index.tsx
@@ -4,20 +4,26 @@ export default function FlexRow({
   children,
   className,
   gap = 0,
+  justifyContent,
   pb = 0,
+  mt = 0,
   style = {}
 }: {
   children: ReactNode
   className?: string
   gap?: number
+  justifyContent?: string
+  mt?: number
   pb?: number
   style?: CSSProperties
 }) {
-  const defaultStyleObj = {
+  const defaultStyleObj: CSSProperties = {
     alignItems: 'center',
-    display: 'flex',
     columnGap: gap ? `${gap}px` : undefined,
+    display: 'flex',
     flexDirection: 'row',
+    justifyContent: justifyContent ? justifyContent : undefined,
+    marginTop: mt ? `${mt}px` : undefined,
     paddingBottom: pb ? `${pb}px` : undefined,
     width: '100%'
   }